config: Add missing ims apn's for supported IMS carriers
Taken from stock oem partitions.

This is a combination of all supported carriers for lilac, poplar, poplar_canada and poplar_dsds. It should hopefully cover all maple carriers as well.

Also, hide the IMS apn's from the user. There is no need to show them in the list, and they should not be selectable.

Finally, added KPN Mobiel Internet from stock as well (mcc 204 / mnc 69) as it was missing.